+++ extensions/quick-notes/src/components/actions.tsx
@@ -10,6 +10,7 @@ import DeleteTags from "./deleteTags";
 import { useCachedState } from "@raycast/utils";
 import { useResetAtom } from "jotai/utils";
 import slugify from "slugify";
+import path from "path";
 
 const Actions = ({
   noNotes,
@@ -81,7 +82,11 @@ const Actions = ({
               icon={{ source: Icon.Folder, tintColor: getTintColor("turquoise") }}
               shortcut={{ modifiers: ["cmd"], key: "o" }}
               onAction={() => {
-                open(`${preferences.fileLocation}/${slugify(`${title}`)}.md`);
+                if (!preferences.fileLocation) {
+                  showToast({ style: Toast.Style.Failure, title: "No Auto Save Location Set" });
+                  return;
+                }
+                open(path.join(preferences.fileLocation, `${slugify(`${title}`)}.md`));
               }}
             />
           </>